Put team-hello in chat.test.js
